Remove the "signed permutation" stuff.

+++ src_dualdesc/OLD/ConvertNC_to_NC_FB_FF.cpp
@@ -29,7 +29,7 @@ int main(int argc, char *argv[]) {
     //
     using T = mpq_class;
     using Tint = mpz_class;
-    using Tidx = int16_t;
+    using Tidx = uint32_t;
     using Telt = permutalib::DoubleSidedPerm<Tidx>;
     using Tgroup = permutalib::Group<Telt, Tint>;
     using Torbsize = uint16_t;
